Added .info() to Beets, with more detailed information.
Made a possibly ill-advised change to potentials which lets you use even a single potential as an iterator, so that the user doesn't need to know whether a potential is a Potential or a PotentialSum before trying to iterate over it.

Moved simulation diagnostics up to Simulation from the ElectricFieldSimulation subclass.