Some new plotting and animation features.
Added debug message for mask application and modified the normalization warnings.

Made a directory under ionization/science for non-testing local scripts. The first script in there is one to generate movies for sinc pulses.

ElectricFieldWindows are now a subclass of Potential, so they have access to PotentialSum Better naming for RadialCosineMask attributes.